DP-6517: A custom jQuery plugin for alerts
This updates the alerts ajax functionality by allowing it to be
more easily overridden, passing in a transform function that can
be used to alter the data between the endpoint and the render
process.
