+++ styleguide/source/_patterns/03-organisms/by-template/ajax-alerts.twig
@@ -1,5 +1,6 @@
 {% set endpoint = ajaxAlerts.endpoint ?: '' %}
-<div class="ma__ajax-alerts ma__ajax-alerts--no-alerts js-ajax-alerts" data-js-alerts-endpoint="{{ endpoint }}">
+{% set selector = ajaxAlerts.customSelector ?: 'js-ajax-alerts' %}
+<div class="ma__ajax-alerts ma__ajax-alerts--no-alerts {{ selector }}" data-js-alerts-endpoint="{{ endpoint }}">
   <section class="ma__emergency-alerts js-accordion js-emergency-alerts" data-id="">
     <div class="ma__emergency-alerts__header">
       <div class="ma__emergency-alerts__container">
